[refactor] 금액 타입 변경

+++ src/main/java/com/jaeyeonling/lotto/exception/LowMoneyException.java
@@ -2,10 +2,10 @@ package com.jaeyeonling.lotto.exception;
 
 public class LowMoneyException extends ArithmeticException {
 
-    private static final String ERROR_MESSAGE = "%f 원이 부족하여 구매할 수 없습니다. (잔고: %f, 금액: %f)";
+    private static final String ERROR_MESSAGE = "%d 원이 부족하여 구매할 수 없습니다. (잔고: %d, 금액: %d)";
 
-    public LowMoneyException(final double balance,
-                             final double price) {
+    public LowMoneyException(final long balance,
+                             final long price) {
         super(String.format(ERROR_MESSAGE, price - balance, balance, price));
     }
 }
